+++ motion/rotate.c
@@ -92,7 +92,7 @@ int main(int argc, char* argv[]) {
 
 	int err = 0, ret = 0;
 	AVRational r;
-	unsigned long components = 0;
+	uint8_t components = 0;
 	unsigned long widths[4], heights[4];
 	uint64_t nframes;
 	FFColorProperties color_props;
